Refactor ReleaseMetadata, PushContext
Now they use some shared functionality to make it easier for getting rid
of ExecutionEnvironment::Fake.
